LayoutTests/platform/mac/fast/dom/shadow/gc-shadow-expected.txt

This tests that the JavaScript wrapper objects of shadow DOM objects
are not prematurely garbage collected.

FAIL layoutTestController.shadowRoot(ps[0]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[1]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[2]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[3]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[4]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[5]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[6]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[7]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[8]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
FAIL layoutTestController.shadowRoot(ps[9]).tattoo should be I <3 WebKit (of type string). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
